update: handle null url, renaming loadSong to getAudioSource, extract according logic to custome hook
